requirements.txt
coverage==4.5.3
Django==2.1.7
django-crispy-forms==1.7.2
django-filter==2.1.0
django-widget-tweaks==1.4.3
icalendar==4.0.3
Pillow==5.4.1
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
